move common functions to super class
